requirements.txt
Flask==2.1.1
Pillow==8.3.2